feat: responsiveness

+++ lib/widgets/event_listBuilder.dart
@@ -1,5 +1,3 @@
-
-
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:hng_events_app/features/events/edit_event/edit_event.dart';
@@ -18,18 +16,18 @@ class EvenListBuilder extends ConsumerWidget {
   });
   final List<Event> events;
   @override
-  Widget build(BuildContext context, WidgetRef ref) {
+  Widget build(BuildContext ctx, WidgetRef ref) {
     return SliverList(
       delegate: SliverChildBuilderDelegate(
         (context, index) {
           return InkWell(
-            onTap: () =>Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => CommentScreen(event: events[index]),
-                ) ,),
+            onTap: () => Navigator.push(
+              context,
+              MaterialPageRoute(
+                builder: (context) => CommentScreen(event: events[index]),
+              ),
+            ),
             child: TimelineEventCard(
-             
               onEdit: () {
                 Navigator.push(
                     context,
@@ -41,29 +39,64 @@ class EvenListBuilder extends ConsumerWidget {
               },
               eventId: events[index].id,
               onDelete: (eventId) {
-                ref
-                    .read(EventProvider.provider)
-                    .deleteEvent(eventId)
-                    .then((value) => {
-                          if (value)
-                            {
-                              showSnackBar(context, 'Event deleted successfully',
-                                  Colors.green),
-                              ref.refresh(allEventsProviderOld),
-                              ref.refresh(userEventsProvider),
-                              ref.refresh(upcomingEventsProviderOld),
-                              ref.read(GroupProvider.groupProvider).getGroups(),
-                            }
-                          else
-                            {
-                              showSnackBar(
-                                  context,
-                                  ref.read(EventProvider.provider).error,
-                                  Colors.red),
-                            }
-                        });
+                showDialog(
+                    context: ctx,
+                    builder: (context) {
+                      return AlertDialog(
+                        title: const Text("Delete event"),
+                        content: const Text(
+                            "Are you sure you want to delete this event?"),
+                        actions: [
+                          TextButton(
+                            onPressed: () async {
+                              try {
+                                Navigator.of(context).pop();
+                                ref
+                                    .read(EventProvider.provider)
+                                    .deleteEvent(eventId)
+                                    .then((value) {
+                                  if (value) {
+                                    ref.refresh(myEventsProvider);
+                                    ref.refresh(allEventsProvider);
+                                    ref.refresh(upcomingEventsProvider);
+                                    ref.refresh(groupsProvider);
+                                    ref.refresh(EventProvider.provider);
+                                     //ref.read(GroupProvider.groupProvider).getGroups();
+                                    showSnackBar(
+                                        ctx,
+                                        'Event deleted successfully',
+                                        Colors.green);
+                                    ref
+                                        .read(GroupProvider.groupProvider)
+                                        .getGroups();
+                                  } else {
+                                    showSnackBar(
+                                        ctx,
+                                        ref.read(EventProvider.provider).error,
+                                        Colors.red);
+                                  }
+                                });
+                              } catch (e) {
+                                ScaffoldMessenger.of(ctx).showSnackBar(
+                                    const SnackBar(
+                                        content: Text(
+                                            "Event could not be deleted ")));
+                              }
+                            },
+                            child: const Text("Yes"),
+                          ),
+                          TextButton(
+                            onPressed: () {
+                              Navigator.of(context).pop();
+                            },
+                            child: const Text("No"),
+                          ),
+                        ],
+                      );
+                    });
               },
-              showVert: ref.read(appUserProvider)?.id == events[index].creatorId,
+              showVert:
+                  ref.read(appUserProvider)?.id == events[index].creatorId,
               context: context,
               screensize: MediaQuery.of(context).size,
               event: events[index],
@@ -72,10 +105,6 @@ class EvenListBuilder extends ConsumerWidget {
         },
         childCount: events.length,
       ),
-     
     );
   }
 }
-
-
-
